v9.0 update
SMAA edge detection (whether color or luma) has problems with scenes that are very blue or very unsaturated (grayscale). To work around this, HQAA now runs luma edge detection a total of 4 times to construct its edge data.
First pass: uses normal backbuffer as input
Second pass: uses a copy of the backbuffer with the color channels left-shifted
Third pass: uses a copy of the backbuffer with the color channels right-shifted
Fourth pass: uses a copy of the backbuffer with the negative of the color channels

Due to these changes luma edge detection will no longer attempt to fall back to color edge detection, it will simply return no edges if it does not detect any. Since this removes a fair amount of math from each iteration of luma edge detection, performance of this version is not significantly worse than version 8.

Since each new source of backbuffer data requires a copy of the back buffer to be written, the HQAA technique's expected number of passes is now 11 when inspected in the statistics tab.